Move existing unit tests to separate unit test subfolder
